Remove node_diff
